0.19.12 new package for Weapons + refactoring
