+++ wfrp/src/main/java/com/warhammer/wfrp/weapons/WeaponRepository.java
@@ -1,6 +1,5 @@
-package com.warhammer.wfrp.repository.weapon;
+package com.warhammer.wfrp.weapons;
 
-import com.warhammer.wfrp.model.weapon.Weapon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WeaponRepository extends JpaRepository<Weapon, Integer> {
