+++ wfrp/src/main/java/com/warhammer/wfrp/weapons/ammo/AmmoRepository.java
@@ -1,6 +1,5 @@
-package com.warhammer.wfrp.repository.weapon;
+package com.warhammer.wfrp.weapons.ammo;
 
-import com.warhammer.wfrp.model.weapon.Ammo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AmmoRepository extends JpaRepository<Ammo, Integer> {
